change data types

+++ src/configs/shop.cairo
@@ -2,10 +2,10 @@ use starknet::ContractAddress;
 
 #[derive(Copy, Drop, IntrospectPacked, Serde)]
 pub struct ShopConfig {
-    pub traditional_cards_quantity: u16,
-    pub modifiers_cards_quantity: u16,
-    pub specials_cards_quantity: u16,
-    pub loot_boxes_quantity: u16,
-    pub power_ups_quantity: u16,
-    pub poker_hands_quantity: u16,
+    pub traditional_cards_quantity: u32,
+    pub modifiers_cards_quantity: u32,
+    pub specials_cards_quantity: u32,
+    pub loot_boxes_quantity: u32,
+    pub power_ups_quantity: u32,
+    pub poker_hands_quantity: u32,
 }
